// src/app/features/lms/grp/models/categoryDetails.ts
export interface CategoryDetailsDto {
    average_anb: null
    average_earnings_per_member: null
    base_premium: null
    base_sum_assured: null
    category_category: string
    category_rate_division_factor: null
    category_unique_code: number
    fee_amount: null
    multiple_earnings_period: null
    period: number;
    premium: null
    premium_mask_code: number;
    previous_category_code: number;
    quotation_code: number
    rate: null
    school_code: null
    short_description: number
    sum_assured: null
    sum_assured_per_member: null
    total_member_earnings: null
    total_members: null
    total_original_loan_amount: null
    total_students: null
    use_cvr_rate: string
    premium_mask_desc: string
}